divide reviewapp_controller
